+++ examples/minimum/variables.tf
@@ -1,1 +1,55 @@
-#(empty)
+variable "name" {
+  description = "The name of the stack"
+  type        = string
+  default     = "minimum-ec2-example"
+}
+
+variable "supporting_resources_name" {
+  description = "Name of supporting resource VPC"
+  type        = string
+  default     = "terraform-aws-ec2"
+}
+
+variable "instance_type" {
+  description = "The instance type to use for the instance. Updates to this field will trigger a stop/start of the EC2 instance"
+  type        = string
+  default     = "t3.small"
+}
+
+variable "monitoring" {
+  description = "If true, the launched EC2 instance will have detailed monitoring enabled which pulls every 1m and adds additonal cost, default monitoring doesn't add costs"
+  type        = bool
+  default     = true
+}
+
+variable "ebs_optimized" {
+  description = "If true, the launched EC2 instance will be EBS-optimized"
+  type        = bool
+  default     = true
+}
+
+variable "root_block_device" {
+  description = "Configuration block to customize details about the root block device of the instance."
+  type        = list(any)
+  default = [
+    {
+      volume_size = 15
+      encrypted   = true
+    }
+  ]
+}
+
+variable "tags" {
+  description = "Map of tags to assign to the resource. Note that these tags apply to the instance and not block storage devices."
+  type        = map(string)
+  default = {
+    Environment        = "example"
+    "user::CostCenter" = "terraform-registry"
+    InstanceScheduler  = true
+    Department         = "DevOps"
+    Project            = "Examples"
+    Owner              = "Boldlink"
+    LayerName          = "Example"
+    LayerId            = "Example"
+  }
+}